+++ packages/openapi2-parser/CHANGELOG.md
@@ -8,8 +8,7 @@
   does not contain the `swagger` key, in prior versions the parser may have
   crashed under some inputs.
 
-- Improved clarity of validation error messages for Schema Object 'type'
-  property.
+- Improved clarity of validation error messages.
 
 ## 0.32.0 (2020-06-12)